Sync PutDimVector declaration and definition
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@27418 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/util/id_mux.hpp
@@ -135,7 +135,7 @@ public:
     const TDimVector&  GetDimVector(size_t i) const;
 
     /// Get modification access to dimension vector
-    TDimVector& PutDimVector(unsigned i);
+    TDimVector& PutDimVector(size_t i);
 
     /// Resize dimesion, create bitset content
     void InitDim(size_t i, size_t dim_size);
@@ -221,6 +221,9 @@ END_NCBI_SCOPE
 /*
 * ---------------------------------------------------------------------------
 * $Log$
+* Revision 1.2  2006/03/23 13:32:10  gouriano
+* Sync PutDimVector declaration and definition
+*
 * Revision 1.1  2006/03/22 19:30:36  kuznets
 * initial revision
 *
